Validatsiya to'girlandi va Error responselar ham success responselar ham hamma apilarda bir xil shablonda kelsin

+++ app/Http/Controllers/AuthController.php
@@ -2,18 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Helpers\ApiResponse;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Hash;
 
 class AuthController extends Controller
 {
     public function register(Request $request)
     {
-        // 1 savol javob
         if (User::where('phone', $request['phone'])->exists()) {
-            return response()->json([
-                'message' => 'Bu raqam bilan oldin ro‘yxatdan o‘tilgan'
-        ], 409);
+             return ApiResponse::error('Bu raqam bilan oldin ro‘yxatdan o‘tilgan',400);
         }
 
         $fields = $request->validate([
@@ -23,6 +22,7 @@ class AuthController extends Controller
             'password' => 'required|string|min:6',
         ]);
 
+        // 1 savol javob yuqorida
 
         $user = User::create([
             'phone' => $fields['phone'],
@@ -33,10 +33,13 @@ class AuthController extends Controller
 
         $token = $user->createToken('auth_token')->plainTextToken;
 
-        return response()->json([
+        $data = [
             'user' => $user,
             'token' => $token
-        ], 201);
+        ];
+
+        return ApiResponse::success('Siz muvaffaqiyatli Ro\'yhatdan o\'tdingiz',$data, 200);
+        //zadacha data
     }
 
     public function login(Request $request)
@@ -50,21 +53,105 @@ class AuthController extends Controller
 
         $token = $user->createToken('auth_token')->plainTextToken;
 
-        return response()->json([
+        $data = [
             'user' => $user,
             'token' => $token
-        ]);
+        ];
+
+        return ApiResponse::success('Siz tizimga muvaffaqiyatli kirdingiz!', $data, 200);
     }
 
-    public function me(Request $request){
+    public function reset_password(Request $request){
+        $fields = $request->validate([
+            'phone' => 'required|string|min:9',
+            'password' => 'required|string|min:6',
+        ]);
 
-        return $request->user();
+        $user = User::where('phone', $fields['phone'])->first();
+
+        if ( $user ) {
+
+            $user->update([
+                'password' => Hash::make($request['password'])
+            ]);
+
+            $data = [
+                'user' => $user
+            ];
+
+            return ApiResponse::success("Parolingiz muvaffaqiyatli o'\zgartirildi!",$data, 200);
+
+        } else {
+            return ApiResponse::error("User not found", 400);
+        }
+    }
+
+    //Update
+    public function update(Request $request){
+        // dd($request);
+        $fields = $request->validate([
+            'phone' => 'required|unique:users,phone,' .$request->user()->id,
+            'firstname' => 'required|max:255',
+            'lastname' => 'required|max:255',
+            'password' => 'required|min:6'
+        ]);
+
+        $user = User::where('phone', $fields['phone'])->first();
+
+        // dd($user);
+
+        if (!$user){
+            return ApiResponse::error('Bunday telefon raqam bilan foydalanuvchi topilmadi!', 400);
+        }
+
+        $user->update([
+            'phone' => $fields['phone'],
+            'firstname' => $fields['firstname'],
+            'lastname' => $fields['lastname'],
+            'password' => Hash::make($fields['password']),
+        ]);
+
+        $data = [
+            'user' => $user
+        ];
+
+        return ApiResponse::success('Malumotlaringiz muvaffaqiyatli o\'zgartirildi!', $data, 200);
+    }
+    //data zadacha bajarildi
+
+    public function me(Request $request){
+        // dd($request->user()->id);
+        $fields = $request->validate([
+            'phone' => 'required|unique:users,phone,'.$request->user()->id,
+            'password' => 'required|min:6',
+        ]);
+
+        $user = User::where('phone', $fields['phone'])->first();
+        $data = [
+            'user' => $user
+        ];
+        if($user){
+            return ApiResponse::success('Siz muvaffaqiyatli kirdingiz!',$data, 200);
+        }
+        return ApiResponse::error('Telefon raqam yoki Parolda xatolik!', 400);
     }
 
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
+        $fields = $request->validate([
+            'phone' => 'required|unique:users,phone,' .$request->user()->id,
+            'password' => 'required|min:6'
+        ]);
+        $user = User::where('phone', $fields['phone'])->first();
+        // dd($user['phone']);
+        if($user){
 
-        return response()->json(['message' => 'Tizimdan chiqildi']);
+            $request->user()->currentAccessToken()->delete();
+
+            return ApiResponse::success("Tizimdan chiqildi!", 200);
+        }else{
+            return ApiResponse::error('Siz Tizimga kirmagansiz!', 400);
+        }
+
     }
 }
